moving Converter interface to converter package
