+++ vraptor-core/src/main/java/br/com/caelum/vraptor/converter/BigIntegerConverter.java
@@ -23,7 +23,6 @@ import java.math.BigInteger;
 import javax.enterprise.context.ApplicationScoped;
 
 import br.com.caelum.vraptor.Convert;
-import br.com.caelum.vraptor.Converter;
 
 /**
  * VRaptor's BigInteger converter.
