+++ vraptor-core/src/main/java/br/com/caelum/vraptor/converter/Converter.java
@@ -15,7 +15,9 @@
  * limitations under the License. 
  */
 
-package br.com.caelum.vraptor;
+package br.com.caelum.vraptor.converter;
+
+import br.com.caelum.vraptor.TwoWayConverter;
 
 
 /**
