+++ vraptor-core/src/main/java/br/com/caelum/vraptor/converter/LocaleBasedCalendarConverter.java
@@ -29,7 +29,6 @@ import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
 import br.com.caelum.vraptor.Convert;
-import br.com.caelum.vraptor.Converter;
 
 /**
  * Locale based calendar converter.
