+++ vraptor-core/src/main/java/br/com/caelum/vraptor/converter/LocaleBasedDateConverter.java
@@ -27,7 +27,6 @@ import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
 import br.com.caelum.vraptor.Convert;
-import br.com.caelum.vraptor.Converter;
 
 /**
  * Locale based date converter.
