+++ vraptor-core/src/main/java/br/com/caelum/vraptor/converter/LocaleBasedPrimitiveFloatConverter.java
@@ -28,7 +28,6 @@ import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 
 import br.com.caelum.vraptor.Convert;
-import br.com.caelum.vraptor.Converter;
 
 /**
  * Localized version of VRaptor's Float converter. If the input value if empty or a null string, null value is returned.
