+++ vraptor-core/src/main/java/br/com/caelum/vraptor/converter/jodatime/LocalDateTimeConverter.java
@@ -28,9 +28,9 @@ import org.joda.time.format.DateTimeFormat;
 import org.joda.time.format.DateTimeFormatter;
 
 import br.com.caelum.vraptor.Convert;
-import br.com.caelum.vraptor.Converter;
 import br.com.caelum.vraptor.converter.ConversionException;
 import br.com.caelum.vraptor.converter.ConversionMessage;
+import br.com.caelum.vraptor.converter.Converter;
 
 /**
  * VRaptor converter for {@link LocalDateTime}. {@link LocalDateTime} is part of Joda Time library.
