+++ vraptor-core/src/test/java/br/com/caelum/vraptor/ioc/fixture/ConverterInTheClasspath.java
@@ -17,7 +17,7 @@
 package br.com.caelum.vraptor.ioc.fixture;
 
 import br.com.caelum.vraptor.Convert;
-import br.com.caelum.vraptor.Converter;
+import br.com.caelum.vraptor.converter.Converter;
 
 @Convert(Void.class)
 public class ConverterInTheClasspath implements Converter<Void> {
